+++ crates/stackable-versioned-macros/src/lib.rs
@@ -454,8 +454,11 @@ pub struct FooSpec {
     bar: usize,
     baz: bool,
 }
+
+# fn main() {
 let merged_crd = Foo::merged_crd("v1").unwrap();
 println!("{}", serde_yaml::to_string(&merged_crd).unwrap());
+# }
 ```
 "#
 )]